portfolio/requirements.txt
asgiref==3.8.1
boto3==1.37.6
botocore==1.37.6
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
jmespath==1.0.1
Markdown==3.7
pillow==11.1.0
psycopg==3.2.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
